- first part of the e2e tests refactoring
