+++ .github/workflows/slow-tests-1.yml
@@ -1,0 +1,27 @@
+name: Tests
+
+on:
+  push:
+    branches: [ main, feat/* ]
+  pull_request:
+    branches: [ main, feat/* ]
+
+jobs:
+  test:
+    name: Slow-tests-1
+    runs-on: ubuntu-latest
+    steps:
+      - name: Set up Go 1.x
+        uses: actions/setup-go@v3
+        with:
+          go-version: 1.20.7
+        id: go
+
+      - name: Check out code
+        uses: actions/checkout@v3
+
+      - name: Get dependencies
+        run: |
+          go get -v -t -d ./...
+      - name: Slow tests 01
+        run: make slow-tests-01
