+++ integrationTests/relayers/slowTests/refundWithChainSimulator_test.go
@@ -14,7 +14,6 @@ import (
 	bridgeCore "github.com/multiversx/mx-bridge-eth-go/core"
 	"github.com/multiversx/mx-bridge-eth-go/integrationTests/relayers/slowTests/framework"
 	"github.com/multiversx/mx-sdk-go/data"
-	"github.com/stretchr/testify/require"
 )
 
 func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
@@ -45,7 +44,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -82,7 +81,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -93,7 +92,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		)
 	})
 	t.Run("unknown function should refund", func(t *testing.T) {
-		callData := createScCallData("unknownFunction", 50000000)
+		callData := CreateScCallData("unknownFunction", 50000000)
 		usdcToken := GenerateTestUSDCToken()
 		usdcToken.TestOperations[2].MvxSCCallData = callData
 		usdcToken.TestOperations[2].MvxFaultySCCall = true
@@ -119,7 +118,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -130,7 +129,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		)
 	})
 	t.Run("not a valid function name should refund", func(t *testing.T) {
-		callData := createScCallData("=", 50000000)
+		callData := CreateScCallData("=", 50000000)
 		usdcToken := GenerateTestUSDCToken()
 		usdcToken.TestOperations[2].MvxSCCallData = callData
 		usdcToken.TestOperations[2].MvxFaultySCCall = true
@@ -156,7 +155,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -167,7 +166,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		)
 	})
 	t.Run("empty function with no args should refund", func(t *testing.T) {
-		callData := createScCallData("", 50000000)
+		callData := CreateScCallData("", 50000000)
 		usdcToken := GenerateTestUSDCToken()
 		usdcToken.TestOperations[2].MvxSCCallData = callData
 		usdcToken.TestOperations[2].MvxFaultySCCall = true
@@ -193,7 +192,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -205,7 +204,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 	})
 	t.Run("empty function with args should refund", func(t *testing.T) {
 		dummyAddress := strings.Repeat("2", 32)
-		callData := createScCallData("", 50000000, dummyAddress)
+		callData := CreateScCallData("", 50000000, dummyAddress)
 		usdcToken := GenerateTestUSDCToken()
 		usdcToken.TestOperations[2].MvxSCCallData = callData
 		usdcToken.TestOperations[2].MvxFaultySCCall = true
@@ -231,7 +230,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -242,7 +241,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		)
 	})
 	t.Run("uninitialized contract should refund", func(t *testing.T) {
-		callData := createScCallData("claim", 50000000)
+		callData := CreateScCallData("claim", 50000000)
 		uninitializedSCAddressBytes, _ := data.NewAddressFromBech32String("erd1qqqqqqqqqqqqqpgqcc69ts8409p3h77q5chsaqz57y6hugvc4fvs64k74v")
 		usdcToken := GenerateTestUSDCToken()
 		usdcToken.TestOperations[2].MvxSCCallData = callData
@@ -274,7 +273,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].InvalidReceiver = uninitializedSCAddressBytes.AddressBytes()
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -285,7 +284,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		)
 	})
 	t.Run("built-in function should refund", func(t *testing.T) {
-		callData := createScCallData("SaveKeyValue", 50000000, "6b657930", "76616c756530")
+		callData := CreateScCallData("SaveKeyValue", 50000000, "6b657930", "76616c756530")
 		usdcToken := GenerateTestUSDCToken()
 		usdcToken.TestOperations[2].MvxSCCallData = callData
 		usdcToken.TestOperations[2].MvxFaultySCCall = true
@@ -311,7 +310,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -352,7 +351,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxForceSCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -363,7 +362,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		)
 	})
 	t.Run("0 gas limit should refund", func(t *testing.T) {
-		callData := createScCallData("callPayable", 0)
+		callData := CreateScCallData("callPayable", 0)
 		usdcToken := GenerateTestUSDCToken()
 		usdcToken.TestOperations[2].MvxSCCallData = callData
 		usdcToken.TestOperations[2].MvxFaultySCCall = true
@@ -389,7 +388,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -400,7 +399,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		)
 	})
 	t.Run("small gas limit should refund", func(t *testing.T) {
-		callData := createScCallData("callPayable", 2000)
+		callData := CreateScCallData("callPayable", 2000)
 		usdcToken := GenerateTestUSDCToken()
 		usdcToken.TestOperations[2].MvxSCCallData = callData
 		usdcToken.TestOperations[2].MvxFaultySCCall = true
@@ -426,7 +425,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -437,7 +436,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		)
 	})
 	t.Run("high gas limit should refund", func(t *testing.T) {
-		callData := createScCallData("callPayable", 610000000)
+		callData := CreateScCallData("callPayable", 610000000)
 		usdcToken := GenerateTestUSDCToken()
 		usdcToken.TestOperations[2].MvxSCCallData = callData
 		usdcToken.TestOperations[2].MvxFaultySCCall = true
@@ -463,7 +462,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -474,7 +473,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		)
 	})
 	t.Run("extra parameter should refund", func(t *testing.T) {
-		callData := createScCallData("callPayable", 50000000, "extra parameter")
+		callData := CreateScCallData("callPayable", 50000000, "extra parameter")
 		usdcToken := GenerateTestUSDCToken()
 		usdcToken.TestOperations[2].MvxSCCallData = callData
 		usdcToken.TestOperations[2].MvxFaultySCCall = true
@@ -500,7 +499,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -511,7 +510,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		)
 	})
 	t.Run("no arguments should refund", func(t *testing.T) {
-		callData := createScCallData("callPayableWithParams", 50000000)
+		callData := CreateScCallData("callPayableWithParams", 50000000)
 		usdcToken := GenerateTestUSDCToken()
 		usdcToken.TestOperations[2].MvxSCCallData = callData
 		usdcToken.TestOperations[2].MvxFaultySCCall = true
@@ -537,7 +536,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -548,7 +547,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		)
 	})
 	t.Run("wrong number of arguments should refund", func(t *testing.T) {
-		callData := createScCallData("callPayableWithParams", 50000000, string([]byte{37}))
+		callData := CreateScCallData("callPayableWithParams", 50000000, string([]byte{37}))
 		usdcToken := GenerateTestUSDCToken()
 		usdcToken.TestOperations[2].MvxSCCallData = callData
 		usdcToken.TestOperations[2].MvxFaultySCCall = true
@@ -574,7 +573,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -588,7 +587,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		malformedUint64String := string([]byte{37, 36, 35, 34, 33, 32, 31, 32, 33}) // 9 bytes instead of 8
 		dummyAddress := strings.Repeat("2", 32)
 
-		callData := createScCallData("callPayableWithParams", 50000000, malformedUint64String, dummyAddress)
+		callData := CreateScCallData("callPayableWithParams", 50000000, malformedUint64String, dummyAddress)
 		usdcToken := GenerateTestUSDCToken()
 		usdcToken.TestOperations[2].MvxSCCallData = callData
 		usdcToken.TestOperations[2].MvxFaultySCCall = true
@@ -614,7 +613,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -625,7 +624,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		)
 	})
 	t.Run("wrong arguments encoding should refund", func(t *testing.T) {
-		callData := createScCallData("callPayableWithParams", 50000000)
+		callData := CreateScCallData("callPayableWithParams", 50000000)
 		// the last byte is the data missing marker, we will replace that
 		callData[len(callData)-1] = bridgeCore.DataPresentProtocolMarker
 		// add garbage data
@@ -656,7 +655,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -670,7 +669,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		dummyAddress := strings.Repeat("2", 32)
 		dummyUint64 := string([]byte{37})
 
-		callData := createScCallData("callPayableWithParams", 50000000, dummyUint64, dummyAddress)
+		callData := CreateScCallData("callPayableWithParams", 50000000, dummyUint64, dummyAddress)
 		callData[3] += 60 // we simulate that the buffer was longer, but it was trimmed
 
 		usdcToken := GenerateTestUSDCToken()
@@ -698,7 +697,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		mexToken.TestOperations[2].MvxFaultySCCall = true
 		ApplyMEXRefundBalances(&mexToken)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -712,7 +711,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		ethereumLimit := 130727
 		buff := make([]byte, ethereumLimit)
 		_, _ = rand.Read(buff)
-		callData := createScCallData("callPayableWithBuff", 100000000, string(buff))
+		callData := CreateScCallData("callPayableWithBuff", 100000000, string(buff))
 
 		usdcToken := GenerateTestUSDCToken()
 		usdcToken.TestOperations = []framework.TokenOperations{
@@ -788,14 +787,14 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		}
 		usdcToken.SpecialChecks.WrapperDeltaLiquidityCheck = big.NewInt(0)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
 		)
 	})
 	t.Run("not a valid function name and small gas limit should refund", func(t *testing.T) {
-		callData := createScCallData("=", 1)
+		callData := CreateScCallData("=", 1)
 
 		usdcToken := GenerateTestUSDCToken()
 		usdcToken.TestOperations = []framework.TokenOperations{
@@ -871,7 +870,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 		}
 		usdcToken.SpecialChecks.WrapperDeltaLiquidityCheck = big.NewInt(0)
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -1145,7 +1144,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 			{
 				ValueToTransferToMvx: big.NewInt(1500),
 				ValueToSendFromMvX:   nil,
-				MvxSCCallData:        createScCallData("callPayable", 50000000),
+				MvxSCCallData:        CreateScCallData("callPayable", 50000000),
 			},
 		}
 		eurocToken.DeltaBalances = map[framework.HalfBridgeIdentifier]framework.DeltaBalancesOnKeys{
@@ -1303,7 +1302,7 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 			WrapperDeltaLiquidityCheck: big.NewInt(0),
 		}
 
-		testRelayersWithChainSimulatorAndTokensAndRefund(
+		NewTestEnvironmentWithChainSimulatorAndTokensAndRefund(
 			t,
 			make(chan error),
 			usdcToken,
@@ -1313,40 +1312,4 @@ func TestRelayersShouldExecuteTransfersWithRefund(t *testing.T) {
 			mexToken,
 		)
 	})
-}
-
-func testRelayersWithChainSimulatorAndTokensAndRefund(tb testing.TB, manualStopChan chan error, tokens ...framework.TestTokenParams) {
-	flows := createFlowsBasedOnToken(tb, tokens...)
-
-	setupFunc := func(tb testing.TB, setup *framework.TestSetup) {
-		for _, flow := range flows {
-			flow.setup = setup
-		}
-
-		setup.IssueAndConfigureTokens(tokens...)
-		setup.MultiversxHandler.CheckForZeroBalanceOnReceivers(setup.Ctx, tokens...)
-		for _, flow := range flows {
-			flow.handlerToStartFirstBridge(flow)
-		}
-	}
-
-	processFunc := func(tb testing.TB, setup *framework.TestSetup) bool {
-		allFlowsFinished := true
-		for _, flow := range flows {
-			allFlowsFinished = allFlowsFinished && flow.process()
-		}
-
-		// commit blocks in order to execute incoming txs from relayers
-		setup.EthereumHandler.SimulatedChain.Commit()
-		setup.ChainSimulator.GenerateBlocks(setup.Ctx, 1)
-		require.LessOrEqual(tb, setup.ScCallerModuleInstance.GetNumSentTransaction(), setup.GetNumScCallsOperations())
-
-		return allFlowsFinished
-	}
-
-	_ = testRelayersWithChainSimulator(tb,
-		setupFunc,
-		processFunc,
-		manualStopChan,
-	)
 }
